client-influx-db/requirements.txt
asyncua==1.1.0
influxdb-client==1.43.0
python-dotenv
